+++ lib/config.dart
@@ -12,6 +12,6 @@ class InFluxConfig {
   //content:
   static final String rssFeedUrl = 'http://www.tagesschau.de/xml/rss2';
 
-  static final String youtubeApiKey = "API-KEY-GOES-HERE";
+  static final String youtubeApiKey = "AIzaSyC8FMty17OjZdNnCKD3cYf9lrt7O4EJ7XA";
   static final String youtubeChannelId = 'UCb5TfGtSgvNPVPQawfCFuAw';
 }
