+++ lib/utility/Youtube/model/youtube_channel_info.dart
@@ -1,4 +1,4 @@
-import 'package:influx/utility/Youtube/thumbnail_size.dart';
+import 'package:influx/utility/Youtube/model/thumbnail_size.dart';
 import 'package:meta/meta.dart';
 
 class YoutubeChannelInfo {
