+++ lib/utility/Youtube/model/youtube_channel_with_videos.dart
@@ -1,5 +1,5 @@
-import 'package:influx/utility/Youtube/youtube_channel_info.dart';
-import 'package:influx/utility/Youtube/youtube_video_info.dart';
+import 'package:influx/utility/Youtube/model/youtube_channel_info.dart';
+import 'package:influx/utility/Youtube/model/youtube_video_info.dart';
 import 'package:meta/meta.dart';
 
 class YoutubeChannelWithVideos{
