+++ test/Youtube/youtubeApiAdapter_getVideos_test.dart
@@ -1,8 +1,8 @@
 import 'package:http/http.dart';
 import 'package:influx/config.dart';
-import 'package:influx/utility/Youtube/thumbnail_size.dart';
+import 'package:influx/utility/Youtube/model/thumbnail_size.dart';
 import 'package:influx/utility/Youtube/youtube_api_adapter.dart';
-import 'package:influx/utility/Youtube/youtube_video_info.dart';
+import 'package:influx/utility/Youtube/model/youtube_video_info.dart';
 import 'package:test_api/test_api.dart';
 
 import 'http_client_youtube_wbs_mock.dart';
